+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * stringId base can be customized in an input box
 * New function: **Get String Contents**
   Gathers all stringIds and correcponding texts from a mission file and copies them for further usage in the CNC translations table
+* Support for multitple windows (gamepad windows) in **Create Tutorial StringIds**
 
 ### 0.0.2